requirements.txt
colorlog==6.4.1
logger==1.4
Pillow==8.3.
